order backed done

+++ src/main/java/com/increff/employee/service/OrderItemService.java
@@ -1,5 +1,6 @@
 package com.increff.employee.service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -58,7 +59,7 @@ public class OrderItemService {
 
 
     @Transactional(rollbackOn = ApiException.class)
-    public void deleteOrder(Integer orderId) {
+    public void deleteOrder(Integer orderId) throws ApiException {
         deleteOrderJust(orderId);
         List<OrderItemPojo> orderItemPojoList = dao.selectAll();
         List<InventoryPojo> inventoryPojoList = inventoryService.getAll();
@@ -82,28 +83,41 @@ public class OrderItemService {
     @Transactional(rollbackOn = ApiException.class)
     public void delete(Integer id) throws ApiException {
         OrderItemPojo orderItemPojo = dao.select(id);
-        List<OrderItemPojo> orderItemPojoList = dao.selectAll();
-        int count=0;
-        for(int i=0;i<orderItemPojoList.size();i++){
-            if(orderItemPojoList.get(i).getOrderId()==orderItemPojo.getOrderId()) {
-                count ++;
+        OrderPojo orderPojo = orderService.getOrderById(orderItemPojo.getOrderId());
+        if(orderPojo.isEditable()==true) {
+            List<OrderItemPojo> orderItemPojoList = dao.selectAll();
+            int count = 0;
+            for (int i = 0; i < orderItemPojoList.size(); i++) {
+                if (orderItemPojoList.get(i).getOrderId() == orderItemPojo.getOrderId()) {
+                    count++;
+                }
             }
+            if (count < 2) {
+                //orderService.delete(orderItemPojo.getOrderId());
+                deleteOrderJust(orderItemPojo.getOrderId());
+            }
+            //psuedo inventory pojo
+            InventoryPojo inventoryPojo = inventoryService.getFromProductId(orderItemPojo.getProductId());
+            Integer quantity = inventoryPojo.getQuantity();
+            quantity += orderItemPojo.getQuantity();
+            inventoryPojo.setQuantity(quantity);
+            orderPojo.setTime(LocalDateTime.now());
+            dao.delete(id);
         }
-        if(count<2) {
-            //orderService.delete(orderItemPojo.getOrderId());
-            deleteOrderJust(orderItemPojo.getOrderId());
+        else {
+            throw new ApiException("this order is not editable");
         }
-        //psuedo inventory pojo
-        InventoryPojo inventoryPojo  = inventoryService.getFromProductId(orderItemPojo.getProductId());
-        Integer quantity = inventoryPojo.getQuantity();
-        quantity += orderItemPojo.getQuantity();
-        inventoryPojo.setQuantity(quantity);
-        dao.delete(id);
     }
 
     @Transactional(rollbackOn = ApiException.class)
-    public void deleteOrderJust(Integer orderId) {
-        orderService.delete(orderId);
+    public void deleteOrderJust(Integer orderId) throws ApiException {
+        OrderPojo orderPojo = orderService.getOrderById(orderId);
+        if(orderPojo.isEditable()==true) {
+            orderService.delete(orderId);
+        }
+        else {
+            throw new ApiException("this order is not editable");
+        }
     }
 
     @Transactional(rollbackOn = ApiException.class)
@@ -116,9 +130,11 @@ public class OrderItemService {
         OrderPojo orderPojo = orderService.getOrderById(orderItemPojo1.getOrderId());
         //OrderItemPojo orderItemPojo = dao.select(orderItemPojo1.getProductId());
         if(orderPojo.isEditable()==true){
+            getInventoryFromProductId(orderItemPojo1,orderItemPojo1.getId());
             OrderItemPojo orderItemPojo = getCheck(id);
             orderItemPojo.setQuantity(orderItemPojo1.getQuantity());
             orderItemPojo.setSellingPrice(orderItemPojo1.getSellingPrice());
+            orderPojo.setTime(LocalDateTime.now());
             if(orderItemPojo1.getQuantity()==0) {
                 delete(orderItemPojo1.getId());
             }
@@ -180,20 +196,20 @@ public class OrderItemService {
     }
 
     @Transactional(rollbackOn = ApiException.class)
-    public void getInventoryFromProductId(OrderItemEditForm orderItemEditForm, Integer id) throws ApiException {
+    public void getInventoryFromProductId(OrderItemPojo orderItemEditPojo, Integer id) throws ApiException {
         OrderItemPojo orderItemPojo = dao.select(id);
-        InventoryPojo inventoryPojo = inventoryService.getFromProductId(orderItemEditForm.getProductId());
+        InventoryPojo inventoryPojo = inventoryService.getFromProductId(orderItemEditPojo.getProductId());
         ProductPojo productPojo = productService.get(inventoryPojo.getProductId());
         if(inventoryPojo==null) {
             throw new ApiException("this product is not present in the inventory");
         }
-        if(orderItemEditForm.getQuantity()>inventoryPojo.getQuantity()+orderItemPojo.getQuantity()){
+        if(orderItemEditPojo.getQuantity()>inventoryPojo.getQuantity()+orderItemPojo.getQuantity()){
             throw new ApiException("quantity of item more than the quantity in inventory");
         }
-        if(orderItemEditForm.getSelling_price()>productPojo.getMrp()){
+        if(orderItemEditPojo.getSellingPrice()>productPojo.getMrp()){
             throw new ApiException("price of item can't be more than MRP");
         }
-        inventoryPojo.setQuantity((inventoryPojo.getQuantity()+orderItemPojo.getQuantity()-orderItemEditForm.getQuantity()));
+        inventoryPojo.setQuantity((inventoryPojo.getQuantity()+orderItemPojo.getQuantity()-orderItemEditPojo.getQuantity()));
     }
 
     @Transactional
